Minor fixups
